migrate GetConfigurations() to parser package
